Change BeeTV's link and fix spelling in "Cunrchyroll mod"
BeeTV has a link to a weird crappy guide on how to download it, I download and installed the app and in it's "About" section it had this URL: https://beetvapk.app/ and it looks much more slick, official and nice so I think that this is the right thing. "Cunrchyroll mod" has a spelling mistake, fixed it to "Crunchyroll mod".